Splitting out types for readability
Splitting out types to their own folder/file for readability and future work.

Updating readme so the import example points to denos hosted version of our module instead.
