+++ types/ZonosCartItem.ts
@@ -1,0 +1,22 @@
+type ZonosCartItem = {
+  countryOfOrigin?: string;
+  description: string;
+  height?: string | number;
+  imageURL: string;
+  itemBrand?: string;
+  itemCategory?: string;
+  itemCustomization?: string;
+  itemHSCode?: string;
+  itemUrl?: string;
+  length?: string | number;
+  nonShippable?: boolean;
+  productId?: string | number;
+  quantity: string | number;
+  sku?: string;
+  unitPrice: string | number;
+  weight?: string | number;
+  weightUnit?: string;
+  width?: string | number;
+};
+
+export type { ZonosCartItem };
