+++ types/ZonosConfig.ts
@@ -1,0 +1,9 @@
+type ZonosConfig = {
+  account_number: string | number;
+  api_key: string;
+  apiUrl?: string;
+  apiVersion?: string | number;
+  corsProxy?: string;
+};
+
+export type { ZonosConfig };
